+++ app/src/main/java/paging/wrapper/model/data/PagingRequest.java
@@ -5,14 +5,14 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class PagingRequest {
 
-  public abstract long offSet();
+  public abstract int offSet();
 
   public abstract PagingQueryParam pagingQueryParam();
 
   public abstract PagingRemoteRequestConfig queryConfig();
 
   public static PagingRequest create(
-      long offSet, PagingQueryParam pagingQueryParam, PagingRemoteRequestConfig queryConfig) {
+      int offSet, PagingQueryParam pagingQueryParam, PagingRemoteRequestConfig queryConfig) {
     return new AutoValue_PagingRequest(offSet, pagingQueryParam, queryConfig);
   }
 }
